Layout and Content changes
This update simplifies the layout (removing the home button), fixes a
formatting error in the display of the {{ COPYRIGHT }} field, added a
{{ CREATOR }} field to acknowledge the contributors of annotations and
field overviews more visibly, changes the contents of the “help” page
(e.g. index.html), adds questions to frame the field overview in
literature, and removes the incomplete articles from the dropdown.